requirements.txt
numpy
Pillow
matplotlib
PySimpleGUI
